Refactor Wikidata search into focused modules and centralize gap cases.
Split read.ts search logic into reusable pieces and share registry, gap-case data, and classification helpers across tests and stats scripts.
