+++ src/lib/wikidata/read.ts
@@ -1,20 +1,10 @@
-import {
-  Entity,
-  EntityId,
-  ItemId,
-  SearchResponse,
-  SearchResult,
-} from 'wikibase-sdk'
+import { Entity, EntityId, ItemId } from 'wikibase-sdk'
 import { Claim, transformFromWikidataDateStringToDate, wbk } from './util'
 import { WbGetEntitiesResponse } from 'wikibase-sdk/dist/src/helpers/parse_responses'
-import { SearchEntitiesOptions } from 'wikibase-sdk/dist/src/queries/search_entities'
 import wikidataConfig from '../../config/wikidata'
-import { lookupKnownCompanyWikidataId } from './knownCompanyLookup'
+import { fetchJsonWithRetries } from './wikidataHttp'
 
-/** `wbgetentities` returns sitelinks; the SDK `Entity` type omits them. */
-type EntityWithSitelinks = Entity & {
-  sitelinks?: Record<string, { title?: string } | undefined>
-}
+export { searchCompany, type CompanySearchResult } from './searchCompany'
 
 const {
   CARBON_FOOTPRINT,
@@ -25,57 +15,6 @@ const {
   APPLIES_TO_PART,
   ARCHIVE_URL,
 } = wikidataConfig.properties
-
-async function fetchJsonWithRetries<T = unknown>(
-  url: string,
-  {
-    headers,
-    maxAttempts = 3,
-    expectedContentType = 'application/json',
-    context,
-  }: {
-    headers?: Record<string, string>
-    maxAttempts?: number
-    expectedContentType?: string
-    context?: string
-  }
-): Promise<T> {
-  let attempt = 0
-  let res: Response | undefined
-  while (attempt < maxAttempts) {
-    res = await fetch(url, { headers })
-    if (res.ok) break
-    if ([429, 502, 503, 504].includes(res.status)) {
-      await new Promise((r) => setTimeout(r, (attempt + 1) * 1000))
-      attempt++
-      continue
-    }
-    break
-  }
-
-  const ctx = context ? `${context} ` : ''
-
-  if (!res) {
-    throw new Error(`${ctx}no response received`)
-  }
-
-  if (!res.ok) {
-    const text = await res.text().catch(() => '')
-    throw new Error(
-      `${ctx}HTTP ${res.status} ${res.statusText} – body: ${text.slice(0, 300)}`
-    )
-  }
-
-  const ct = (res.headers.get('content-type') || '').toLowerCase()
-  if (!ct.includes(expectedContentType)) {
-    const text = await res.text().catch(() => '')
-    throw new Error(
-      `${ctx}returned non-JSON (${ct}) – body: ${text.slice(0, 300)}`
-    )
-  }
-
-  return (await res.json()) as T
-}
 
 export async function getWikipediaTitle(id: EntityId): Promise<string> {
   const url = wbk.getEntities({
@@ -122,909 +61,6 @@ export async function getLEINumber(
   }
 
   return claims['P1278'][0].mainsnak.datavalue.value
-}
-
-/** Trailing company legal forms often omitted from Wikidata labels (e.g. "… AB" vs "…"). */
-const LEGAL_FORM_SUFFIXES = new Set([
-  'ab',
-  'aktiebolag',
-  'ag',
-  'asa',
-  'bv',
-  'corp',
-  'corporation',
-  'gmbh',
-  'inc',
-  'incorporated',
-  'int',
-  'international',
-  'koncern',
-  'limited',
-  'llc',
-  'llp',
-  'lp',
-  'ltd',
-  'nv',
-  'oy',
-  'oyj',
-  'plc',
-  'publ',
-  'sa',
-  'as',
-  'scandinavia',
-  'company',
-])
-
-function normalizeCompanySearchName(companyName: string): string {
-  return companyName.replace(/,/g, ' ').trim().replace(/\s+/g, ' ')
-}
-
-/**
- * Removes one or more trailing legal-form tokens (e.g. AB, Ltd, or "(AB)", "(publ)").
- * Returns null if nothing was removed or the remainder would be empty.
- */
-function stripLegalFormSuffixes(companyName: string): string | null {
-  let s = normalizeCompanySearchName(companyName)
-  if (!s) return null
-  const original = s
-
-  while (s.length > 0) {
-    const parenMatch = s.match(/^(.+?)\s*\(([^)]+)\)\s*$/i)
-    if (parenMatch) {
-      const rest = parenMatch[1].trim().replace(/^\.+|\.+$/g, '')
-      const innerNorm = parenMatch[2].toLowerCase().replace(/\./g, '').trim()
-      if (rest && LEGAL_FORM_SUFFIXES.has(innerNorm)) {
-        s = rest
-        continue
-      }
-    }
-
-    const match = s.match(/^(.+?)\s+([^\s]+)\.?$/i)
-    if (!match) break
-    const rest = match[1].trim().replace(/^\.+|\.+$/g, '')
-    const lastNorm = match[2].toLowerCase().replace(/\./g, '')
-    if (!LEGAL_FORM_SUFFIXES.has(lastNorm)) break
-    if (!rest) break
-    s = rest
-  }
-
-  return s !== original && s.length > 0 ? s : null
-}
-
-/**
- * Expands the common corporate abbreviation "Int." ("International") so
- * Wikidata search can match labels like "Millicom" that omit the short form
- * (e.g. "Millicom Int. Cellular" → no hits; "Millicom International Cellular" → Q276345).
- */
-function expandInternationalAbbreviation(companyName: string): string | null {
-  const normalized = normalizeCompanySearchName(companyName)
-  if (!/\bInt\./i.test(normalized)) return null
-  const expanded = normalized
-    .replace(/\bInt\.\s*/gi, 'International ')
-    .replace(/\bInt\./gi, 'International ')
-    .trim()
-    .replace(/\s+/g, ' ')
-  return expanded !== normalized ? expanded : null
-}
-
-/**
- * Many Wikidata items use a trailing “Group” while exchange listings drop it.
- * Returns null when the name already ends with “Group” or is empty.
- */
-function appendGroupSuffixForSearch(companyName: string): string | null {
-  const s = normalizeCompanySearchName(companyName)
-  if (!s) return null
-  if (/\bgroup$/i.test(s)) return null
-  return `${s} Group`
-}
-
-function appendAbSuffixForSearch(companyName: string): string | null {
-  const n = normalizeCompanySearchName(companyName)
-  if (!n) return null
-  if (nameAlreadyHasTrailingLegalSuffix(n, 'ab')) return null
-  return `${n} AB`
-}
-
-/**
- * Removes a trailing “Group” token (exchange listings; not in {@link LEGAL_FORM_SUFFIXES}).
- */
-function stripTrailingGroupSuffix(companyName: string): string | null {
-  const s = normalizeCompanySearchName(companyName)
-  if (!s) return null
-  const match = s.match(/^(.+?)\s+group\.?$/i)
-  if (!match) return null
-  const rest = match[1].trim().replace(/^\.+|\.+$/g, '')
-  return rest.length > 0 ? rest : null
-}
-
-/** Baltic / Nordic listing prefixes often omitted from Wikidata labels (e.g. “AB Akola Group”). */
-function stripLeadingLegalPrefix(companyName: string): string | null {
-  const s = normalizeCompanySearchName(companyName)
-  const match = s.match(/^(AB|AS|APB)\s+(.+)$/i)
-  if (!match) return null
-  const rest = match[2].trim()
-  return rest.length > 0 ? rest : null
-}
-
-/**
- * Trailing listing qualifiers Wikidata often drops (e.g. “Bactiguard Holding B” → “Bactiguard”).
- */
-function stripTrailingListingQualifiers(companyName: string): string[] {
-  const n = normalizeCompanySearchName(companyName)
-  if (!n) return []
-  const patterns = [
-    /^(.+?)\s+holding\s+b$/i,
-    /^(.+?)\s+holding$/i,
-    /^(.+?)\s+sweden$/i,
-    /^(.+?)\s+industri$/i,
-    /^(.+?)\s+operations\s+o[üu]$/i,
-    /^(.+?)\s+logistic\s+property$/i,
-  ]
-  const out: string[] = []
-  for (const pattern of patterns) {
-    const match = n.match(pattern)
-    if (match?.[1]?.trim()) out.push(match[1].trim())
-  }
-  return out
-}
-
-/** “I.A.R Systems” → “IAR Systems”. */
-function normalizeInitialsAndPunctuation(companyName: string): string | null {
-  const n = normalizeCompanySearchName(companyName)
-  const simplified = n
-    .replace(/([A-Za-zÅÄÖÜÕŽ])\.\s*/g, '$1')
-    .replace(/\s+/g, ' ')
-    .trim()
-  return simplified !== n && simplified.length > 0 ? simplified : null
-}
-
-/** “XANO Industri” → “Xano Industri” (Wikidata label casing). */
-function titleCaseLeadingAllCapsToken(companyName: string): string | null {
-  const parts = normalizeCompanySearchName(companyName).split(' ')
-  if (parts.length === 0) return null
-  const first = parts[0]
-  if (!/^[A-ZÅÄÖÜÕŽ]{2,}$/.test(first)) return null
-  parts[0] = first.charAt(0) + first.slice(1).toLowerCase()
-  const out = parts.join(' ')
-  return out !== companyName ? out : null
-}
-
-function looksBalticListing(text: string): boolean {
-  return (
-    /[žūüõėą]/i.test(text) ||
-    /^(AB|AS|APB)\s/i.test(text) ||
-    /\bOÜ\b/i.test(text) ||
-    /\bOperations\s+OÜ\b/i.test(text)
-  )
-}
-
-function pushUniqueProbe(candidates: string[], query: string | null | undefined) {
-  if (!query?.trim()) return
-  candidates.push(query.trim())
-}
-
-/**
- * Shorter or transformed search strings when exchange listings diverge from Wikidata labels.
- * Shortest variant first for probe priority.
- */
-function listingStripProbeQueries(companyName: string): string[] {
-  const normalized = normalizeCompanySearchName(companyName)
-  if (!normalized) return []
-
-  const candidates: string[] = []
-  const withoutAb = stripLegalFormSuffixes(companyName)
-  const withoutGroup = stripTrailingGroupSuffix(companyName)
-  const withoutPrefix = stripLeadingLegalPrefix(companyName)
-
-  pushUniqueProbe(candidates, withoutAb)
-  pushUniqueProbe(candidates, withoutGroup)
-  pushUniqueProbe(candidates, withoutPrefix)
-  for (const q of stripTrailingListingQualifiers(companyName)) {
-    pushUniqueProbe(candidates, q)
-  }
-
-  if (withoutAb) {
-    pushUniqueProbe(candidates, stripTrailingGroupSuffix(withoutAb))
-    for (const q of stripTrailingListingQualifiers(withoutAb)) {
-      pushUniqueProbe(candidates, q)
-    }
-  }
-  if (withoutGroup) {
-    pushUniqueProbe(candidates, stripLegalFormSuffixes(withoutGroup))
-    for (const q of stripTrailingListingQualifiers(withoutGroup)) {
-      pushUniqueProbe(candidates, q)
-    }
-  }
-  if (withoutPrefix) {
-    pushUniqueProbe(candidates, stripTrailingGroupSuffix(withoutPrefix))
-    pushUniqueProbe(candidates, stripLegalFormSuffixes(withoutPrefix))
-    for (const q of stripTrailingListingQualifiers(withoutPrefix)) {
-      pushUniqueProbe(candidates, q)
-    }
-  }
-
-  if (/\bi\s+sverige\b/i.test(normalized)) {
-    pushUniqueProbe(
-      candidates,
-      normalized.replace(/\bi\s+sverige\b/i, 'Sverige')
-    )
-    pushUniqueProbe(
-      candidates,
-      normalized.replace(/\bi\s+sverige\b/i, 'Sverige AB')
-    )
-  }
-
-  const dePunctuated = normalizeInitialsAndPunctuation(normalized)
-  pushUniqueProbe(candidates, dePunctuated)
-  if (dePunctuated) {
-    pushUniqueProbe(candidates, stripTrailingGroupSuffix(dePunctuated))
-  }
-
-  pushUniqueProbe(candidates, titleCaseLeadingAllCapsToken(normalized))
-
-  const derived = [...candidates]
-  for (const q of derived) {
-    pushUniqueProbe(candidates, titleCaseLeadingAllCapsToken(q))
-    for (const shorter of stripTrailingListingQualifiers(q)) {
-      pushUniqueProbe(candidates, shorter)
-      pushUniqueProbe(candidates, titleCaseLeadingAllCapsToken(shorter))
-    }
-  }
-
-  if (normalized.split(/\s+/).length === 1) {
-    pushUniqueProbe(candidates, `Investment AB ${normalized}`)
-  }
-
-  const normKey = normalizedHitLabel(normalized)
-  return [...new Set(candidates)]
-    .filter((q) => normalizedHitLabel(q) !== normKey)
-    .sort((a, b) => a.length - b.length)
-    .slice(0, 12)
-}
-
-function listingNameHasStripProbes(companyName: string): boolean {
-  return listingStripProbeQueries(companyName).length > 0
-}
-
-async function mergeListingStripProbeResults(
-  results: SearchResult[],
-  companyName: string,
-  language: SearchEntitiesOptions['language'] | undefined
-): Promise<SearchResult[]> {
-  let merged = results
-  for (const query of listingStripProbeQueries(companyName)) {
-    if (resultsIncludeExactLabel(merged, query)) break
-    const fromProbe = await searchWikidataEntitiesForCompany(query, language, {
-      includeExtraNordicLanguages: looksBalticListing(query),
-    })
-    if (fromProbe.length === 0) continue
-    merged = interleaveDedupeSearchResults(fromProbe, merged)
-    if (resultsIncludeExactLabel(merged, query)) break
-  }
-  return merged
-}
-
-function normalizedHitLabel(label: string | undefined): string {
-  return (label ?? '').trim().toLowerCase().replace(/\s+/g, ' ')
-}
-
-function resultsIncludeExactLabel(
-  results: SearchResult[],
-  expectedLabel: string
-): boolean {
-  const want = normalizedHitLabel(expectedLabel)
-  return results.some((r) => normalizedHitLabel(r.label) === want)
-}
-
-/**
- * Exchange listings often omit “Group” / “AB” while Wikidata labels include them
- * (e.g. listing “Nelly” → item “Nelly Group”). Merge probe hits even when the plain
- * name already returns noisy results.
- */
-function listingSuffixProbeQueries(companyName: string): string[] {
-  const probes: string[] = []
-  const withGroup = appendGroupSuffixForSearch(companyName)
-  if (withGroup) probes.push(withGroup)
-  const withAb = appendAbSuffixForSearch(companyName)
-  if (withAb) probes.push(withAb)
-  return probes
-}
-
-async function mergeListingSuffixProbeResults(
-  results: SearchResult[],
-  companyName: string,
-  language: SearchEntitiesOptions['language'] | undefined
-): Promise<SearchResult[]> {
-  let merged = results
-  for (const query of listingSuffixProbeQueries(companyName)) {
-    if (resultsIncludeExactLabel(merged, query)) continue
-    const fromProbe = await searchWikidataEntitiesForCompany(query, language)
-    if (fromProbe.length === 0) continue
-    merged = interleaveDedupeSearchResults(fromProbe, merged)
-  }
-  return merged
-}
-
-/**
- * Wikidata often lists companies with a legal-form suffix ("Evolution AB", "Acme Ltd").
- * Plain-name search can miss the item entirely (e.g. "Evolution" → biology). When every
- * hit lacks a company-like P31, we probe {@link LEGAL_FORM_SUFFIXES} in order.
- */
-function legalFormSuffixAppendLabel(norm: string): string {
-  const lowerShort: Record<string, string> = {
-    ltd: 'Ltd',
-    plc: 'plc',
-    corp: 'Corp',
-    inc: 'Inc',
-    gmbh: 'GmbH',
-    group: 'Group',
-    publ: '(publ)',
-    int: 'Int',
-  }
-  if (lowerShort[norm]) return lowerShort[norm]
-  if (norm.length >= 5) return norm.charAt(0).toUpperCase() + norm.slice(1)
-  return norm.toUpperCase()
-}
-
-function nameAlreadyHasTrailingLegalSuffix(
-  normalizedName: string,
-  norm: string
-): boolean {
-  const paren = normalizedName.match(/^(.+?)\s*\(([^)]+)\)\s*$/i)
-  if (paren) {
-    const inner = paren[2].toLowerCase().replace(/\./g, '').trim()
-    if (inner === norm) return true
-  }
-  const match = normalizedName.match(/^(.+?)\s+([^\s]+)\.?$/i)
-  if (!match) return false
-  const lastNorm = match[2].toLowerCase().replace(/\./g, '')
-  return lastNorm === norm
-}
-
-/** Priority order: common Stockholm-list style first, then rest A–z. */
-function legalFormNormsForSupplementarySearch(): string[] {
-  return [...LEGAL_FORM_SUFFIXES].sort((a, b) => {
-    if (a === 'ab') return -1
-    if (b === 'ab') return 1
-    return a.localeCompare(b)
-  })
-}
-
-function supplementaryLegalFormListingQueries(companyName: string): string[] {
-  const n = normalizeCompanySearchName(companyName)
-  if (!n) return []
-  const out: string[] = []
-  for (const norm of legalFormNormsForSupplementarySearch()) {
-    if (nameAlreadyHasTrailingLegalSuffix(n, norm)) continue
-    out.push(`${n} ${legalFormSuffixAppendLabel(norm)}`)
-  }
-  return out
-}
-
-const WIKIDATA_SEARCH_HEADERS = {
-  Accept: 'application/json',
-  'User-Agent': 'KlimatkollenGarboBot/1.0 (+https://klimatkollen.se)',
-} as const
-
-/** Industry (P452) — item-valued; exposed on {@link searchCompany} hits under `statements.P452`. */
-const INDUSTRY = 'P452' as const
-
-/**
- * instance of (P31) object ids we treat as business/company-like for ranking.
- * Subclasses are not expanded (no subclass-of traversal).
- */
-const COMPANY_LIKE_INSTANCE_OF = new Set<string>([
-  'Q4830453', // business
-  'Q783794', // company
-  'Q6881511', // enterprise
-  'Q891723', // public company
-  'Q431289', // brand
-  'Q4387609', // architectural firm
-  'Q68295960', // Swedish government agency
-])
-
-function collectInstanceOfIds(entity: Entity | undefined): Set<string> {
-  const ids = new Set<string>()
-  if (!entity || entity.type !== 'item') return ids
-  const p31 = entity.claims?.P31
-  if (!p31) return ids
-  for (const claim of p31) {
-    if (claim.mainsnak.snaktype !== 'value' || !claim.mainsnak.datavalue) {
-      continue
-    }
-    const dv = claim.mainsnak.datavalue
-    if (dv.type === 'wikibase-entityid' && 'id' in dv.value) {
-      ids.add(dv.value.id)
-    }
-  }
-  return ids
-}
-
-function collectItemIdsForProperty(
-  entity: Entity | undefined,
-  property: string
-): ItemId[] {
-  const out: ItemId[] = []
-  if (!entity || entity.type !== 'item') return out
-  const claims = entity.claims?.[property]
-  if (!claims) return out
-  for (const claim of claims) {
-    if (claim.mainsnak.snaktype !== 'value' || !claim.mainsnak.datavalue) {
-      continue
-    }
-    const dv = claim.mainsnak.datavalue
-    if (dv.type === 'wikibase-entityid' && 'id' in dv.value) {
-      out.push(dv.value.id as ItemId)
-    }
-  }
-  return out
-}
-
-/** Country (P17) → Wikipedia site id for ranking with default search language `en`. */
-const P17_COUNTRY_TO_WIKI_SITE: Readonly<Record<string, string>> = {
-  Q34: 'svwiki', // Sweden
-  Q33: 'fiwiki', // Finland
-  Q20: 'nowiki', // Norway
-  Q35: 'dawiki', // Denmark
-  Q183: 'dewiki', // Germany
-  Q142: 'frwiki', // France
-  Q145: 'enwiki', // United Kingdom
-  Q30: 'enwiki', // United States
-  Q16: 'enwiki', // Canada
-  Q408: 'nlwiki', // Netherlands
-  Q38: 'itwiki', // Italy
-  Q29: 'eswiki', // Spain
-  Q211: 'lvwiki', // Latvia
-  Q191: 'eewiki', // Estonia
-  Q37: 'ltwiki', // Lithuania
-  Q36: 'huwiki', // Hungary
-  Q43: 'trwiki', // Turkey
-  Q148: 'zhwiki', // China
-  Q17: 'jawiki', // Japan
-  Q884: 'kowiki', // South Korea
-}
-
-const WIKI_SITES_EXCLUDED_FROM_NOTABILITY_COUNT = new Set([
-  'commonswiki',
-  'specieswiki',
-])
-
-function collectP17CountryIds(entity: Entity | undefined): string[] {
-  const out: string[] = []
-  if (!entity || entity.type !== 'item') return out
-  const p17 = entity.claims?.P17
-  if (!p17) return out
-  for (const claim of p17) {
-    if (claim.mainsnak.snaktype !== 'value' || !claim.mainsnak.datavalue) {
-      continue
-    }
-    const dv = claim.mainsnak.datavalue
-    if (dv.type === 'wikibase-entityid' && 'id' in dv.value) {
-      out.push(dv.value.id)
-    }
-  }
-  return out
-}
-
-function countEncyclopediaSitelinks(
-  sitelinks: EntityWithSitelinks['sitelinks'] | undefined
-): number {
-  if (!sitelinks) return 0
-  let n = 0
-  for (const key of Object.keys(sitelinks)) {
-    if (!key.endsWith('wiki')) continue
-    if (WIKI_SITES_EXCLUDED_FROM_NOTABILITY_COUNT.has(key)) continue
-    const e = sitelinks[key]
-    if (
-      e &&
-      typeof e === 'object' &&
-      'title' in e &&
-      (e as { title?: string }).title
-    )
-      n++
-  }
-  return n
-}
-
-/**
- * Higher = stronger disambiguation signal among company-like hits.
- * Tier gaps keep raw encyclopedia link counts below country/wiki tiers.
- */
-function computeSitelinkPreferenceRank(
-  entity: Entity | undefined,
-  language: SearchEntitiesOptions['language'] | undefined
-): number {
-  if (!entity || entity.type !== 'item') return 0
-  const sl = (entity as EntityWithSitelinks).sitelinks
-  if (!sl) return 0
-
-  const hasSite = (site: string) => {
-    const e = sl[site]
-    return Boolean(
-      e &&
-        typeof e === 'object' &&
-        'title' in e &&
-        (e as { title?: string }).title
-    )
-  }
-
-  if (language && hasSite(`${language}wiki`)) return 1_000_000
-  if ((language === 'nb' || language === 'nn') && hasSite('nowiki'))
-    return 950_000
-
-  for (const c of collectP17CountryIds(entity)) {
-    const site = P17_COUNTRY_TO_WIKI_SITE[c]
-    if (site && hasSite(site)) return 800_000
-  }
-
-  if (hasSite('enwiki')) return 600_000
-
-  return countEncyclopediaSitelinks(sl)
-}
-
-/**
- * Wikidata search returns an item `label` per hit (often distinct from aliases).
- * Prefer labels close to the query so "Mips" → **Mips AB** (Q109787297) ranks above
- * **MIPS Technologies** (Q1631366), which only shares the first token.
- */
-function searchLabelAffinityScore(
-  result: SearchResult,
-  companyName: string
-): number {
-  const q = normalizeCompanySearchName(companyName).toLowerCase()
-  const raw = (result.label ?? '').trim()
-  if (!q || !raw) return 0
-  const label = raw.toLowerCase().replace(/\s+/g, ' ')
-
-  if (label === q) return 500_000
-  if (label === `${q} ab`) return 450_000
-  if (label === `${q} group`) return 450_000
-
-  for (const probe of listingStripProbeQueries(companyName)) {
-    if (label === normalizedHitLabel(probe)) return 460_000
-  }
-
-  if (label === `investment ab ${q}`) return 460_000
-  if (label === `${q} (sweden)`) return 440_000
-  if (label.startsWith(`${q} sverige`)) return 450_000
-
-  const legalSecondToken = new Set([
-    'ab',
-    'ltd',
-    'plc',
-    'corp',
-    'inc',
-    'nv',
-    'ag',
-    'asa',
-    'oy',
-    'bv',
-    'gmbh',
-    'llc',
-    'oyj',
-    'lp',
-    'llp',
-  ])
-  const parts = label.split(/\s+/)
-  if (parts.length === 2 && parts[0] === q) {
-    const tail = parts[1].replace(/\./g, '')
-    if (legalSecondToken.has(tail)) return 440_000
-    return 85_000
-  }
-  if (parts.length >= 2 && parts[0] === q) return 45_000
-
-  return 0
-}
-
-type SearchHitAugmentation = {
-  companyLike: boolean
-  industryIds: ItemId[]
-  sitelinkPreferenceRank: number
-}
-
-async function fetchSearchHitAugmentation(
-  ids: string[],
-  language: SearchEntitiesOptions['language'] | undefined
-): Promise<Map<string, SearchHitAugmentation>> {
-  const byId = new Map<string, SearchHitAugmentation>()
-  if (ids.length === 0) return byId
-
-  const chunkSize = 50
-  for (let offset = 0; offset < ids.length; offset += chunkSize) {
-    const chunk = ids.slice(offset, offset + chunkSize)
-    const url = wbk.getEntities({
-      ids: chunk as EntityId[],
-      props: ['claims', 'sitelinks'],
-      languages: ['en'],
-    })
-    const { entities } = await fetchJsonWithRetries<WbGetEntitiesResponse>(
-      url,
-      {
-        headers: { ...WIKIDATA_SEARCH_HEADERS },
-        maxAttempts: 3,
-        expectedContentType: 'application/json',
-        context: 'Wikidata entities (P31, P452, sitelinks)',
-      }
-    )
-
-    for (const id of chunk) {
-      const entity = entities[id]
-      const instanceOf = collectInstanceOfIds(entity)
-      const companyLike = [...instanceOf].some((q) =>
-        COMPANY_LIKE_INSTANCE_OF.has(q)
-      )
-      const industryIds = collectItemIdsForProperty(entity, INDUSTRY)
-      const sitelinkPreferenceRank = computeSitelinkPreferenceRank(
-        entity,
-        language
-      )
-      byId.set(id, { companyLike, industryIds, sitelinkPreferenceRank })
-    }
-  }
-  return byId
-}
-
-function prioritizeCompanyLikeSearchResults(
-  results: SearchResult[],
-  augmentation: Map<string, SearchHitAugmentation>,
-  companyName: string
-): SearchResult[] {
-  const isCo = (r: SearchResult) => augmentation.get(r.id)?.companyLike === true
-  const rankOf = (r: SearchResult) =>
-    (augmentation.get(r.id)?.sitelinkPreferenceRank ?? 0) +
-    searchLabelAffinityScore(r, companyName)
-
-  const index = new Map(results.map((r, i) => [r.id, i]))
-  const companyHits = results.filter(isCo).sort((a, b) => {
-    const d = rankOf(b) - rankOf(a)
-    if (d !== 0) return d
-    return (index.get(a.id) ?? 0) - (index.get(b.id) ?? 0)
-  })
-  const nonCompany = results.filter((r) => !isCo(r))
-  return [...companyHits, ...nonCompany]
-}
-
-export type CompanySearchResult = SearchResult & {
-  /** Item QIDs for industry (P452), when present on the entity. */
-  statements?: { P452?: ItemId[] }
-}
-
-async function searchWikidataEntities(
-  search: string,
-  language: SearchEntitiesOptions['language'] | undefined
-): Promise<SearchResponse['search']> {
-  const url = wbk.searchEntities({
-    search,
-    type: 'item',
-    language,
-    limit: 50,
-  })
-  const response = await fetchJsonWithRetries<SearchResponse>(url, {
-    headers: { ...WIKIDATA_SEARCH_HEADERS },
-    maxAttempts: 3,
-    expectedContentType: 'application/json',
-    context: 'Wikidata search',
-  })
-  if (response.error) {
-    const msg = response.error.info || response.error.code
-    throw new Error(`Wikidata search failed: ${msg}`)
-  }
-  return response.search ?? []
-}
-
-/**
- * Merges two hit lists by alternating indices and skipping duplicate ids — keeps
- * English crawl order while pulling in Swedish ranking (where Nordic listings
- * such as MTG → Q378944 or SCA → Q52601 often rank #1).
- */
-function interleaveDedupeSearchResults(
-  primary: SearchResult[],
-  secondary: SearchResult[]
-): SearchResult[] {
-  const seen = new Set<string>()
-  const out: SearchResult[] = []
-  const n = Math.max(primary.length, secondary.length)
-  for (let i = 0; i < n; i++) {
-    if (i < primary.length) {
-      const r = primary[i]
-      if (!seen.has(r.id)) {
-        seen.add(r.id)
-        out.push(r)
-      }
-    }
-    if (i < secondary.length) {
-      const r = secondary[i]
-      if (!seen.has(r.id)) {
-        seen.add(r.id)
-        out.push(r)
-      }
-    }
-  }
-  return out
-}
-
-const EXTRA_NORDIC_SEARCH_LANGUAGES: SearchEntitiesOptions['language'][] = [
-  'fi',
-  'et',
-  'lv',
-  'lt',
-]
-
-/**
- * When the UI language is English, also search in Swedish and interleave hits so
- * short tickers and Swedish labels (MTG, SCA, …) are not lost to unrelated en matches.
- * Optionally adds fi/et/lv/lt for Baltic listings or empty-result recovery.
- */
-async function searchWikidataEntitiesForCompany(
-  search: string,
-  language: SearchEntitiesOptions['language'] | undefined,
-  options?: { includeExtraNordicLanguages?: boolean }
-): Promise<SearchResponse['search']> {
-  const primary = await searchWikidataEntities(search, language)
-  let merged = primary
-  if (language === 'en') {
-    const svHits = await searchWikidataEntities(search, 'sv')
-    merged = interleaveDedupeSearchResults(merged, svHits)
-  }
-  if (options?.includeExtraNordicLanguages) {
-    for (const lang of EXTRA_NORDIC_SEARCH_LANGUAGES) {
-      if (lang === language) continue
-      const hits = await searchWikidataEntities(search, lang)
-      if (hits.length > 0) {
-        merged = interleaveDedupeSearchResults(merged, hits)
-      }
-    }
-  }
-  return merged
-}
-
-function knownCompanySearchResult(
-  companyName: string,
-  wikidataId: string
-): CompanySearchResult[] {
-  return [
-    {
-      id: wikidataId,
-      label: companyName,
-      description: '',
-      url: `https://www.wikidata.org/wiki/${wikidataId}`,
-      match: { type: 'label', language: 'en', text: companyName },
-    },
-  ]
-}
-
-export async function searchCompany({
-  companyName,
-  /** `wbsearchentities` UI language; default `en` suits international company names. Pass `sv`, `de`, etc. when you know the listing’s primary locale. */
-  language = 'en',
-  /**
-   * When true, return a Klimatkollen-known Wikidata id from the bundled registry
-   * or Garbo API before calling Wikidata search. Off by default so search-quality
-   * tests still exercise live `wbsearchentities`.
-   */
-  useKnownIdLookup = false,
-}: {
-  companyName: string
-  language?: SearchEntitiesOptions['language']
-  useKnownIdLookup?: boolean
-}): Promise<CompanySearchResult[]> {
-  if (useKnownIdLookup) {
-    const knownId = await lookupKnownCompanyWikidataId(companyName)
-    if (knownId) return knownCompanySearchResult(companyName, knownId)
-  }
-
-  const balticListing = looksBalticListing(companyName)
-  let results = await searchWikidataEntitiesForCompany(companyName, language, {
-    includeExtraNordicLanguages: balticListing,
-  })
-
-  if (results.length === 0) {
-    const simplified = stripLegalFormSuffixes(companyName)
-    if (simplified) {
-      results = await searchWikidataEntitiesForCompany(simplified, language, {
-        includeExtraNordicLanguages: balticListing,
-      })
-    }
-  }
-
-  if (results.length === 0 && companyName.includes('Int')) {
-    const expanded = expandInternationalAbbreviation(companyName)
-    if (expanded) {
-      results = await searchWikidataEntitiesForCompany(expanded, language, {
-        includeExtraNordicLanguages: balticListing,
-      })
-    }
-  }
-
-  if (listingNameHasStripProbes(companyName)) {
-    results = await mergeListingStripProbeResults(
-      results,
-      companyName,
-      language
-    )
-  }
-
-  results = await mergeListingSuffixProbeResults(
-    results,
-    companyName,
-    language
-  )
-
-  if (results.length > 0) {
-    let augmentation = await fetchSearchHitAugmentation(
-      results.map((r) => r.id),
-      language
-    )
-
-    /** e.g. "Lundin Mining Corp." → only legal case Q137125375; stripped "Lundin Mining" finds Q1537901 */
-    const everyHitNonCompany = results.every(
-      (r) => !augmentation.get(r.id)?.companyLike
-    )
-    const hasCompanyLikeListingMatch = results.some(
-      (r) =>
-        augmentation.get(r.id)?.companyLike === true &&
-        searchLabelAffinityScore(r, companyName) >= 440_000
-    )
-    if (everyHitNonCompany || !hasCompanyLikeListingMatch) {
-      const simplified = stripLegalFormSuffixes(companyName)
-      if (simplified) {
-        const fromStrip = await searchWikidataEntitiesForCompany(
-          simplified,
-          language
-        )
-        if (fromStrip.length > 0) {
-          results = fromStrip
-          augmentation = await fetchSearchHitAugmentation(
-            results.map((r) => r.id),
-            language
-          )
-        }
-      }
-
-      const stillEveryNonCompany = results.every(
-        (r) => !augmentation.get(r.id)?.companyLike
-      )
-      if (stillEveryNonCompany) {
-        for (const listingQuery of supplementaryLegalFormListingQueries(
-          companyName
-        )) {
-          const fromLegal = await searchWikidataEntitiesForCompany(
-            listingQuery,
-            language
-          )
-          if (fromLegal.length === 0) continue
-          const augLegal = await fetchSearchHitAugmentation(
-            fromLegal.map((r) => r.id),
-            language
-          )
-          const anyCompany = fromLegal.some(
-            (r) => augLegal.get(r.id)?.companyLike
-          )
-          if (anyCompany) {
-            results = fromLegal
-            augmentation = augLegal
-            break
-          }
-        }
-      }
-    }
-
-    results = prioritizeCompanyLikeSearchResults(
-      results,
-      augmentation,
-      companyName
-    )
-    results = results.map((r) => {
-      const industryIds = augmentation.get(r.id)?.industryIds ?? []
-      const withIndustry: CompanySearchResult = { ...r }
-      if (industryIds.length > 0) {
-        withIndustry.statements = { P452: industryIds }
-      }
-      return withIndustry
-    })
-  }
-
-  return results
 }
 
 export async function getWikidataEntities(ids: `Q${number}`[]) {
