+++ tests/wikidata/wikidata-search-company-public.spec.ts
@@ -1,12 +1,8 @@
-import type { NamedWikidataCase } from './wikidata-search-helpers'
 import { defineWikidataSearchTagSpec } from './wikidata-search-describe-by-tag'
-
-const PUBLIC_SEARCH_EMPTY_RESULTS: ReadonlyArray<NamedWikidataCase> = [
-  { companyName: 'HKFoods', klimatkollenWikidataId: 'Q139591851' },
-]
+import { WIKIDATA_SEARCH_GAP_CASES } from './gapCases'
 
 defineWikidataSearchTagSpec({
   tag: 'public',
   tagLabel: 'public',
-  emptyResults: PUBLIC_SEARCH_EMPTY_RESULTS,
+  ...WIKIDATA_SEARCH_GAP_CASES.public,
 })
